requirements.txt
# --- Core Backend ---
fastapi[all]==0.111.0             # Compatible with Pydantic v2
uvicorn[standard]==0.29.0         # ASGI server for FastAPI
python-dotenv==1.0.1              # Load env vars from .env

# --- Authentication ---
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4             # Password hashing
python-multipart==0.0.9            # Form data parsing for auth

# --- Database ---
SQLAlchemy==2.0.28                # Modern async-compatible ORM
asyncpg==0.29.0                   # Async PostgreSQL driver
alembic==1.13.1                   # Database migration tool
psycopg2-binary==2.9.9            # For Alembic CLI (non-async)
greenlet==3.0.3                   # Required for async SQLAlchemy

# --- Langchain + LLM (Fully Compatible Stack) ---
langchain==0.3.26                 # Compatible with langchain-core 0.3.66
langchain-core==0.3.69            # Required base
langchain-community==0.3.27       # Community tools
langgraph==0.5.3                  # LangGraph orchestration
openai===1.90.0                  # OpenAI Python SDK
langchain-openai==0.3.28

# --- Qdrant Vector Store ---
qdrant-client==1.15.0             # Qdrant vector DB client

# --- Langchain Helpers / Vector Stores ---
tiktoken==0.7.0                   # Token counting for OpenAI models
httpx==0.27.0                     # Async HTTP client used across services

# Optional vector DBs (commented out because you're using Qdrant)
# faiss-cpu==1.8.0                # FAISS for local in-memory vector search
# pinecone-client==3.2.2         # Pinecone vector DB
# chromadb==0.5.0                # ChromaDB vector store
# weaviate-client==4.5.4         # Weaviate vector DB client

# --- Google Translate ---
google-cloud-translate==3.15.2

# --- OCR ---
pytesseract==0.3.10               # Wrapper for Tesseract OCR
Pillow==10.3.0                    # Image processing lib

# --- Compatibility ---
pydantic==2.11.7                  # Explicit version if needed

# --- Extras ---
duckduckgo-search==8.1.1          # Web search via DuckDuckGo
